agents/requirements.txt
langgraph>=0.2.0
langfuse>=2.0.0
boto3>=1.34.0
httpx>=0.27.0
kafka-python>=2.0.2
psycopg[binary]>=3.2.2
pyyaml>=6.0.1
